requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
pandas==2.1.4
fuzzywuzzy==0.18.0
python-Levenshtein           # optional speed-up for fuzzywuzzy
groq==0.4.2
httpx<0.26
plotly==5.22.0
passlib[bcrypt]==1.7.4
